Separated Pytocs.Core tests and fixes *.csproj files
