+++ src/Cli/Pytocs.Cli.csproj
@@ -1,14 +1,14 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFrameworks>net471;netcoreapp2.1</TargetFrameworks>
     <OutputType>exe</OutputType>
-    <AssemblyName>pytocs</AssemblyName>
+    <AssemblyName>Pytocs.Cli</AssemblyName>
     <RootNamespace>Pytocs.Cli</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Core\Core.csproj" />
+    <ProjectReference Include="..\Core\Pytocs.csproj" />
   </ItemGroup>
 
 </Project>
