+++ src/Core/TypeInference/FakeFileSystem.cs
@@ -15,7 +15,6 @@
 #endregion
 
 #if DEBUG
-using Xunit;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -269,24 +268,6 @@ namespace Pytocs.TypeInference
         public Stream CreateFileStream(string outputFileName, FileMode mode, FileAccess access)
         {
             throw new NotImplementedException();
-        }
-    }
-
-    public class FakeFileSystemTests
-    {
-        [Fact]
-        public void Filesys_Add()
-        {
-            var fs = new FakeFileSystem()
-                .Dir("foo")
-                    .File("foo.py", @"print('Howdy')" + "\r\n")
-                    .File("bar.py", @"def foo():\r\n    pass\r\n")
-                .End()
-                .Dir("end")
-                .End();
-            Assert.True(fs.FileExists("\\foo\\foo.py"));
-            Assert.False(fs.FileExists("\\foo\\foo1.py"));
-            Assert.False(fs.FileExists("\\bar\\foo1.py"));
         }
     }
 }
