+++ src/Pytocs.Tests/Translate/LocalVariableGeneratorTests.cs
