+++ src/Pytocs.Tests/TypeInference/AnalyzerTests.cs
